# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IPO Alert API",
    "version": "2.0.0",
    "description": "A comprehensive API for Nepal stock market data, IPO tracking, portfolio management, and price alerts",
    "contact": {
      "name": "IPO Alert Team",
      "email": "support@ipo-alert.com"
    }
  },
  "servers": [
    {
      "url": "https://ipo-alert-latest.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "TrendingStock": {
        "type": "object",
        "properties": {
          "ticker": {"type": "string"},
          "ticker_name": {"type": "string"},
          "latest_price": {"type": "string"},
          "points_change": {"type": "number"},
          "percentage_change": {"type": "number"},
          "traded_of_mkt_cap": {"type": "number"}
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean", "example": false},
          "error": {"type": "string"}
        }
      }
    }
  },
  "paths": {
...
    "/api/indices/{range}": {
      "get": {
        "summary": "Get market indices data",
        "tags": ["Public Data"],
        "parameters": [
          {
            "name": "range",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["1d", "1m", "3m", "1y", "5y", "all"]
            },
            "description": "Time range for indices data"
          }
        ],
        "responses": {
          "200": {
            "description": "Market indices data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "type": {"type": "string"},
                    "timeRange": {"type": "string"},
                    "data": {"type": "array"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid range parameter",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
...
    "/api/trending-stocks": {
      "get": {
        "summary": "Get trending stocks",
        "tags": ["Public Data"],
        "responses": {
          "200": {
            "description": "List of trending stocks by trading volume",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "type": {"type": "string"},
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/TrendingStock"}
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch trending stocks",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
...
  },
...
}